feat: Add functions for fetching the cards
The fetching function takes in a number of cards, and with the help of
the function for generating random IDs(up to 1025), fetches the number
of random cards by returning a Promise with the "Promise.all" method
which takes in an array of Promises. Then the Promise resolves once all
responses get parsed.